requirements.txt
gtts
pyttsx3
termcolor
prettytable
pillow
numpy
progress